+++ test/smoke/src/areas/positron/dataexplorer/sparklinesTrend.ts
